refactor(app): Move components to src/
