+++ ui/components.json
@@ -7,9 +7,9 @@
 		"baseColor": "slate"
 	},
 	"aliases": {
-		"components": "$lib/components",
+		"components": "$components",
 		"utils": "$lib/utils/tailwind.utils",
-		"ui": "$lib/components/ui",
+		"ui": "components/ui",
 		"hooks": "$lib/hooks"
 	},
 	"typescript": true,
