+++ ui/src/components/ui/form/form-label.svelte
@@ -2,7 +2,7 @@
 	import type { Label as LabelPrimitive } from "bits-ui";
 	import { getFormControl } from "formsnap";
 	import { cn } from "$lib/utils//tailwind.utils.ts";
-	import { Label } from "$lib/components/ui/label/index.js";
+	import { Label } from "$components/ui/label/index.js";
 
 	type $$Props = LabelPrimitive.Props;
