+++ ui/src/components/ui/spinner/spinner.svelte
