+++ ui/src/routes/(user)/user/logout/+page.svelte
@@ -1,0 +1,1 @@
+<!-- nothing -->
